src/main/java/codingbat/map2/AllSwap.java
package codingbat.map2;

import java.util.HashMap;
import java.util.Map;

public class AllSwap {
    public String[] allSwap(String[] strings) {
        Map<String, Integer> map = new HashMap<>();
        for (int i = 0; i < strings.length; i++) {
            String key = strings[i].substring(0, 1);
            if (map.containsKey(key)) {
                String temp = strings[i];
                strings[i] = strings[map.get(key)];
                strings[map.get(key)] = temp;
                map.remove(key);
            } else {
                map.put(strings[i].substring(0, 1), i);
            }

        }
        return strings;
    }
}
